wip: adds filter viewer JavaFX utility application for visualizing filters designed through the filter factory.

+++ src/dsp/filter/GoertzelFilter.java
@@ -32,7 +32,7 @@ public class GoertzelFilter
     private int mSampleRate;
     private long mTargetFrequency;
     private int mBlockSize;
-    private double[] mWindowCoefficients;
+    private float[] mWindowCoefficients;
 
     private double mCoefficient;
